requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
anthropic>=0.96.0
openai>=1.54.0
pypdf==5.1.0
python-dotenv==1.0.1
json_repair>=0.30

